+++ CHANGES.rst
@@ -26,6 +26,13 @@ New Features
   some more attributes for inspection by scikit-learn: ``__sklearn_tags__``,
   ``classes_``, ``_estimator_type``. :pr:`1931` by :user:`Jérôme Dockès
   <jeromedockes>`.
+- It is now possible to pass additional (dynamically computed) arguments to the
+  cross-validation splitter used by :class:`DataOp` objects for validation,
+  hyperparameter search etc. For example, the groups for a
+  :class:`sklearn.model_selection.GroupKFold` can be computed as part of the
+  DataOp evaluation and used for splitting. This is achieved by passing the
+  splitter and its arguments to :meth:`DataOp.skb.mark_as_X`. :pr:`1943` by
+  :user:`Jérôme Dockès <jeromedockes>`.
 - :func:`selectors.has_nulls` now takes a ``proportion`` parameter, which allows
   selecting columns that have a fraction of null values above the given threshold.
   :pr:`1881` by :user:`Gabriela Gómez Jiménez <gabrielapgomezji>`.
@@ -48,9 +55,19 @@ Changes
   the ``SKB_DATA_DIRECTORY`` environment variable. The environment variable ``SKRUB_DATA_DIRECTORY``
   is deprecated and will be removed in a future version of skrub.
   :pr:`1852` by :user:`Riccardo Cappuzzo<rcap107>`.
-  :class:`SingleColumnTransformer` and associated exception :class:`RejectColumn` (used
+- :class:`core.SingleColumnTransformer` and associated exception :class:`core.RejectColumn` (used
   internally by many skrub estimators) have been added to the public API, in the newly-created
   :package:`skrub.core` module. :pr:`1851` by :user:`Eloi Massoulié <emassoulie>`.
+- Added the strings ``"None"`` and ``"none"`` to the list of null string values in
+  :class:`Cleaner`. Also, exposed the list of null string values that will be set
+  to null by the :class:`Cleaner` as the parameter ``null_strings``.
+  :pr:`1952` and :pr:`1954` by :user:`Lisa McBride <lisaleemcb>`.
+- The configuration parameter "use_table_report" has been removed from the skrub
+  configuration. Use :meth:`patch_display` instead.
+  :pr:`1973` by :user:`Riccardo Cappuzzo<rcap107>`.
+- The overplotting of the counts atop the vertical histogram bars in the
+  :class:'TableReport' has been removed due to formatting issues.
+  :pr:`1984` by :user:`Lisa McBride<lisaleemcb>`.
 
 Bug Fixes
 --------
@@ -60,15 +77,22 @@ Bug Fixes
   and :pr:`1900` by :user:`Jérôme Dockès <jeromedockes>`.
 - Errors raised when a polars LazyFrame is passed where an eager DataFrame is
   expected are now clearer. :pr:`1916` by :user:`Jérôme Dockès <jeromedockes>`.
+- :meth:`DataOp.skb.cross_validate` would raise an error when passed
+  ``return_indices=True``. Now it returns the train and test indices of each
+  fold in the ``train_indices`` and ``test_indices`` columns of the result
+  dataframe. :pr:`1953` by :user:`Jérôme Dockès <jeromedockes>`.
 - :class:`CheckInputDataFrame` no longer collects Polars LazyFrames automatically;
   a ``TypeError`` is now raised instead, consistent with the rest of the library.
   :pr:`1941` by :user:`Mudit Atrey <MuditAtrey>`.
+- :func:`fetch_employee_salaries` now correctly writes the train and test
+  split CSV files to their respective paths when ``split`` is specified.
+  :pr:`1964` by :user:`MuditAtrey <MuditAtrey>`.
 
 Documentation
 -------------
 - Updated gallery examples to load datasets from their file paths using
   ``pd.read_csv()``, following the pattern established in :pr:`1852`.
-  :pr:`1940` by :user:`MuditAtrey <MuditAtrey>`.
+  :pr:`1940` and :pr:`1964` by :user:`MuditAtrey <MuditAtrey>`.
 
 Release 0.7.2
 =============
@@ -166,9 +190,6 @@ Changes
 - Added ``cast_to_str`` parameter to :class:`Cleaner` to prevent unintended
   conversion of list/object-like columns to strings unless explicitly enabled.
   :pr:`1789` by :user:`PilliSiddharth`.
-- Added the strings ``"None"`` and ``"none"`` to the list of null string values in
- :class:`Cleaner`.
-  :pr:`1952` by :user:`Lisa McBride <lisaleemcb>`.
 
 Bugfixes
 --------